requirements.txt
Django==2.2
djangorestframework==3.9.2
pytz==2020.1
sqlparse==0.3.1
